+++ src/main/java/duke/Command.java
@@ -1,0 +1,2 @@
+package duke;public class Command {
+}
