+++ src/main/java/duke/EndDuke.java
@@ -1,0 +1,7 @@
+package duke.action;
+
+public class endDuke {
+    public static void run(){
+        
+    }
+}
