+++ src/main/java/duke/RunDuke.java
@@ -1,0 +1,7 @@
+package duke.action;
+
+public class RunDuke {
+    public static void run(){
+
+    }
+}
