+++ src/main/java/duke/StartDuke.java
@@ -1,0 +1,2 @@
+package duke;public class StartDuke {
+}
